refactor(cli): move version command out of root
Appears to be a remainder from early days. This puts it in the same
module as all other commands, while also changing the code structure
to align with recent changes.
